Store Original HTML
Store a copy of the original HTML extracted from the chapter page
alongside a copy that all filtering / transformation has been applied
to.  This way we can prevent future downloads of chapter content,
but still make code changes to filtering of chapter content. Then
we just need to re-apply filtering to the original html content
and overwrite Chapter.html with the result.